+++ src/android/com/outsystems/plugins/healthfitness/store/AdvancedQuery.kt
@@ -37,10 +37,10 @@ private val timeUnits: Map<String, EnumTimeUnit> by lazy {
 data class ProcessedBucket(
     val startDate : Long,
     var endDate : Long,
-    var dataPoints : MutableList<DataPoint>,
-    var processedDataPoints : MutableList<Float>,
-    var DEBUG_startDate : String,
-    var DEBUG_endDate : String
+    var dataPoints : MutableList<DataPoint> = mutableListOf(),
+    var processedDataPoints : MutableList<Float> = mutableListOf(),
+    var DEBUG_startDate : String = "",
+    var DEBUG_endDate : String = ""
 )
 
 class AdvancedQuery(